Removed support for protein inference results from emozi XML schema.
This decision is based on a given protein list only making sense in the context of cutoffs applied to PSMs, peptides or proteins. Since emozi XML supports uploading PSMs, peptides, and proteins at all cutoffs and the cutoffs may be applied within the web app, itself, it doesn't make sense to upload protein inference results in this document.