Removed "default_filter" attribute from filterable annotation types.
New behavior: if a "default_filter_value" attribute is defined for a filterable annotation type, the data will be filtered by default on that annotation type using the supplied value. For example, one may supply a "default_filter_value=0.01" for a q-value annotation type to filter the data by default in the emozi web app on a q-value level of 0.01.